+++ packages/admin/src/Actions/Pricing/UpdateTieredPricing.php
@@ -4,7 +4,7 @@ namespace Lunar\Hub\Actions\Pricing;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Collection;
-use Illuminate\Support\Facades\DB;
+use Lunar\Facades\DB;
 use Lunar\Models\Currency;
 use Lunar\Models\Price;