+++ packages/core/CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 0.2.7
+
+### Fixed
+
+- The `addAddress` method for a `Cart` will now pass the address to the pipelines correctly.
+
+## 0.2.6
+
+### Fixed
+
+- When deleting a discount the users relationship will now `detach` rather than trying to `delete`.
+
 ## 0.2.5
 
 ### Fixed